Small tweaks to edit warehouse page to make the correct post api call, also moved warehouses page into a component and refactored the page using the component
