+++ src/pages/Warehouses/Warehouses.scss
@@ -1,221 +1,23 @@
 @use "../../styles/partials/mixins" as *;
 @use "../../styles/partials/variables" as *;
 
-.head {
-  padding: 1rem;
-  padding-bottom: 1.5rem;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
+.warehouse-list {
+  margin: 16px;
+  margin-top: -4rem;
+  border-radius: 3px;
+  box-shadow: 0px 2px 5px 0px #13182c1a;
+  z-index: 1;
+  background-color: $primary-white;
+
   @include tablet {
-    padding: 1.5rem;
-    flex-direction: row;
-    align-items: center;
+    margin: auto;
+    margin-top: -5rem;
+    max-width: 45rem;
   }
 
   @include desktop {
-    padding: 2rem;
-  }
-  &__header {
-    padding-bottom: 1rem;
-    &--name {
-      @include page-header;
-    }
-
-    @include tablet {
-      padding-bottom: unset;
-    }
-  }
-
-  &__actions {
-    display: flex;
-    flex-direction: column;
-
-    @include tablet {
-      flex-direction: row;
-      align-items: center;
-      gap: 1rem;
-    }
-  }
-
-  &__search {
-    margin-bottom: 1rem;
-    display: flex;
-    justify-content: space-between;
-    border-radius: 20px;
-    border: 1px solid $secondary-cloud;
-    height: 38px;
-    padding: 0rem 1rem;
-    background-color: $primary-white;
-    align-items: center;
-
-    @include tablet {
-      margin: unset;
-      width: 20vw;
-    }
-
-    &--desc {
-      @include body-medium;
-      color: $secondary-slate;
-    }
-  }
-
-  &__add {
-    text-decoration: none;
-    @include primary-button;
-  }
-}
-
-.warehouse {
-  margin-top: -1rem; //to be tested once nav is put in
-  margin: 2rem; //FOR TESTING
-  border-radius: 3px;
-  box-shadow: 0px 2px 5px 0px #13182c1a;
-
-  &__head {
-    padding: 1rem;
-    padding-bottom: 1.5rem;
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-
-    @include tablet {
-      padding: 1.5rem;
-    }
-
-    @include desktop {
-      padding: 2rem;
-    }
-  }
-  &__header {
-    @include page-header;
-  }
-  &__sort-icon {
-    display: none;
-    @include tablet {
-      display: block;
-      height: 14px;
-    }
-  }
-
-  &__row {
-    display: grid;
-    grid-template-columns: repeat(2, 1fr);
-    padding: 16px;
-    border-top: 1px solid $secondary-cloud;
-    @include body-medium;
-    @include tablet {
-      padding: 0px;
-      display: grid;
-      grid-template-columns: repeat(2, 1fr) auto;
-      &:hover {
-        background-color: rgba(46, 102, 229, 0.07);
-      }
-
-      &:nth-of-type(2) {
-        border-top: none;
-        padding-top: 0;
-
-        .warehouse__cell-item {
-          padding-bottom: 8px;
-          height: 50%;
-        }
-      }
-    }
-  }
-  &__column {
-    @include tablet {
-      display: grid;
-      grid-template-columns: repeat(2, 1fr);
-    }
-  }
-  &__row:nth-of-type(n + 3) &__cell-label {
-    @include tablet {
-      display: none;
-    }
-  }
-  &__link {
-    text-decoration: none;
-  }
-  &__cell {
-    padding: 8px 0;
-    @include tablet {
-      padding: 0;
-    }
-    &-item {
-      @include tablet {
-        padding: 16px 4px;
-      }
-      height: 100%;
-      display: flex;
-      align-items: center;
-    }
-    &-label {
-      @include table-header;
-      text-transform: uppercase;
-      color: $secondary-slate;
-      padding-bottom: 4px;
-      @include tablet {
-        padding: 16px 4px;
-        background-color: $background-lightgrey;
-        display: flex;
-        align-items: center;
-        gap: 4px;
-
-        &--first {
-          padding-left: 24px;
-        }
-
-        &--last {
-          padding-right: 24px;
-        }
-      }
-    }
-    &--actions {
-      grid-column: span 2;
-      @include tablet {
-        grid-column: auto;
-      }
-    }
-    &--actions &-label {
-      display: none;
-      @include tablet {
-        display: block;
-        text-align: right;
-      }
-    }
-  }
-  &__name {
-    font-weight: 600;
-    color: $primary-indigo;
-    gap: 4px;
-    @include tablet {
-      padding-left: 24px;
-    }
-    &:hover {
-      text-decoration: underline;
-    }
-  }
-  &__contact-info {
-    flex-direction: column;
-    align-items: start;
-    justify-content: center;
-  }
-  &__phone,
-  &__email {
-    @include tablet {
-      display: flex;
-      justify-content: start;
-    }
-  }
-  &__actions {
-    display: flex;
-    width: stretch;
-    justify-content: space-between;
-    @include tablet {
-      padding-right: 24px;
-      gap: 12px;
-    }
+    margin: auto;
+    margin-top: -5rem;
+    max-width: 70rem;
   }
 }
